+++ yash-env/CHANGELOG.md
@@ -17,7 +17,8 @@ A _private dependency_ is used internally and not visible to downstream users.
     - This module re-exports `Code`, `Location`, `Source`, and `pretty` from
       `yash_syntax::source`.
 - `parser` module
-    - This module contains the `Config` struct for configuring the parser.
+    - `Config`: Configuration for the parser.
+    - `is_name`: Function that checks if a string is a valid variable name.
 - `alias` module
     - This module re-exports `Alias`, `AliasSet`, `Glossary`, and `HashEntry`
       from `yash_syntax::alias`.